thread.py overwrote chrome.py
